# api/swagger/swagger.yaml
definitions:
...
  domain.ProgressStatus:
    enum:
    - new
    - skipped
    - known
    - learned
    type: string
    x-enum-comments:
      ProgressStatusKnown: User marked the question as known
      ProgressStatusLearned: User marked the question as learned
      ProgressStatusNew: User has not seen the question
      ProgressStatusSkipped: User skipped the question
    x-enum-descriptions:
    - User has not seen the question
    - User skipped the question
    - User marked the question as known
    - User marked the question as learned
    x-enum-varnames:
    - ProgressStatusNew
    - ProgressStatusSkipped
    - ProgressStatusKnown
    - ProgressStatusLearned
  domain.UserQuestionProgress:
    properties:
      id:
        type: string
      questionID:
        type: string
      status:
        $ref: '#/definitions/domain.ProgressStatus'
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  progress.aggregateResponse:
    properties:
      current_position:
        type: integer
      known:
        type: integer
      learned:
        type: integer
      new:
        type: integer
      skipped:
        type: integer
    type: object
  progress.markRequest:
    properties:
      question_id:
        type: string
      topic_id:
        type: string
    required:
    - question_id
    - topic_id
    type: object
...
  render.Error:
    properties:
      details: {}
      error:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /progress/known:
    post:
      consumes:
      - application/json
      description: Mark a question as known for the authenticated user
      parameters:
      - description: Progress payload
        in: body
        name: progress
        required: true
        schema:
          $ref: '#/definitions/progress.markRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/render.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Progress not found
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      security:
      - BearerAuth: []
      summary: Mark question as known
      tags:
      - Progress
  /progress/learned:
    post:
      consumes:
      - application/json
      description: Mark a question as learned for the authenticated user
      parameters:
      - description: Progress payload
        in: body
        name: progress
        required: true
        schema:
          $ref: '#/definitions/progress.markRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/render.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Progress not found
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      security:
      - BearerAuth: []
      summary: Mark question as learned
      tags:
      - Progress
  /progress/questions/{id}:
    get:
      description: Get progress of a specific question for the authenticated user
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserQuestionProgress'
        "400":
          description: Invalid question ID
          schema:
            $ref: '#/definitions/render.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Progress not found
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      security:
      - BearerAuth: []
      summary: Get question progress
      tags:
      - Progress
  /progress/skipped:
    post:
      consumes:
      - application/json
      description: Mark a question as skipped for the authenticated user
      parameters:
      - description: Progress payload
        in: body
        name: progress
        required: true
        schema:
          $ref: '#/definitions/progress.markRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/render.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Progress not found
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      security:
      - BearerAuth: []
      summary: Mark question as skipped
      tags:
      - Progress
  /progress/topics/{id}:
    delete:
      description: Reset all progress data for a topic for the authenticated user
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid topic ID
          schema:
            $ref: '#/definitions/render.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Progress not found
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      security:
      - BearerAuth: []
      summary: Reset topic progress
      tags:
      - Progress
    get:
      description: Get aggregated progress for a topic for the authenticated user
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/progress.aggregateResponse'
        "400":
          description: Invalid topic ID
          schema:
            $ref: '#/definitions/render.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Progress not found
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      security:
      - BearerAuth: []
      summary: Get topic progress
      tags:
      - Progress
...
swagger: "2.0"